Relocates solidity source into descrete namespace. blockchain.py -> chains.py
