--- nkms/blockchain/eth/sol_source/contracts/zeppelin/math/Math.sol
@@ -1,24 +1,0 @@
-pragma solidity ^0.4.18;
-
-
-/**
- * @title Math
- * @dev Assorted math operations
- */
-library Math {
-  function max64(uint64 a, uint64 b) internal pure returns (uint64) {
-    return a >= b ? a : b;
-  }
-
-  function min64(uint64 a, uint64 b) internal pure returns (uint64) {
-    return a < b ? a : b;
-  }
-
-  function max256(uint256 a, uint256 b) internal pure returns (uint256) {
-    return a >= b ? a : b;
-  }
-
-  function min256(uint256 a, uint256 b) internal pure returns (uint256) {
-    return a < b ? a : b;
-  }
-}
