--- nkms/blockchain/eth/sol_source/contracts/zeppelin/token/ERC20/DetailedERC20.sol
@@ -1,16 +1,0 @@
-pragma solidity ^0.4.18;
-
-import "./ERC20.sol";
-
-
-contract DetailedERC20 is ERC20 {
-  string public name;
-  string public symbol;
-  uint8 public decimals;
-
-  function DetailedERC20(string _name, string _symbol, uint8 _decimals) public {
-    name = _name;
-    symbol = _symbol;
-    decimals = _decimals;
-  }
-}
