--- nkms/blockchain/eth/sol_source/project.json
@@ -1,142 +1,0 @@
-{
-  "chains": {
-    "mainnetrpc": {
-      "chain": {
-        "class": "populus.chain.external.ExternalChain"
-      },
-      "web3": {
-        "provider": {
-          "class": "web3.providers.rpc.HTTPProvider",
-          "settings": {
-            "endpoint_uri": "http://127.0.0.1:8545",
-            "request_kwargs": {
-                "timeout": 10
-            }
-          }
-        }
-      },
-      "contracts": {
-        "backends": {
-          "JSONFile": {
-            "$ref": "contracts.backends.JSONFile"
-          },
-          "Memory": {
-            "$ref": "contracts.backends.Memory"
-          },
-          "ProjectContracts": {
-            "$ref": "contracts.backends.ProjectContracts"
-          },
-          "TestContracts": {
-            "$ref": "contracts.backends.TestContracts"
-          }
-        }
-      }
-    },
-    "ropsten": {
-      "chain": {
-        "class": "populus.chain.geth.TestnetChain"
-      },
-      "web3": {
-        "$ref": "web3.GethIPC"
-      },
-      "contracts": {
-        "backends": {
-          "JSONFile": {
-            "$ref": "contracts.backends.JSONFile"
-          },
-          "Memory": {
-            "$ref": "contracts.backends.Memory"
-          },
-          "ProjectContracts": {
-            "$ref": "contracts.backends.ProjectContracts"
-          },
-          "TestContracts": {
-            "$ref": "contracts.backends.TestContracts"
-          }
-        }
-      }
-    },
-    "temp": {
-      "chain": {
-        "class": "populus.chain.geth.TemporaryGethChain"
-      },
-      "web3": {
-        "$ref": "web3.GethIPC"
-      },
-      "contracts": {
-        "backends": {
-          "Memory": {
-            "$ref": "contracts.backends.Memory"
-          },
-          "ProjectContracts": {
-            "$ref": "contracts.backends.ProjectContracts"
-          },
-          "TestContracts": {
-            "$ref": "contracts.backends.TestContracts"
-          }
-        }
-      }
-    },
-    "tester": {
-      "chain": {
-        "class": "populus.chain.tester.TesterChain"
-      },
-      "web3": {
-        "$ref": "web3.Tester"
-      },
-      "contracts": {
-        "backends": {
-          "Memory": {
-            "$ref": "contracts.backends.Memory"
-          },
-          "ProjectContracts": {
-            "$ref": "contracts.backends.ProjectContracts"
-          },
-          "TestContracts": {
-            "$ref": "contracts.backends.TestContracts"
-          }
-        }
-      }
-    },
-    "testrpc": {
-      "chain": {
-        "class": "populus.chain.testrpc.TestRPCChain"
-      },
-      "web3": {
-        "$ref": "web3.TestRPC"
-      },
-      "contracts": {
-        "backends": {
-          "Memory": {
-            "$ref": "contracts.backends.Memory"
-          },
-          "ProjectContracts": {
-            "$ref": "contracts.backends.ProjectContracts"
-          },
-          "TestContracts": {
-            "$ref": "contracts.backends.TestContracts"
-          }
-        }
-      }
-    }
-  },
-  "compilation": {
-    "contracts_source_dirs": [
-      "./contracts",
-      "../../tests/contracts"
-    ],
-    "import_remappings": [
-      "contracts=nkms/blockchain/eth/sol_source/contracts"
-    ],
-    "backend": {
-      "settings": {
-        "stdin": {
-          "optimizer": {
-            "enabled": true
-          }
-        }
-      }
-    }
-  },
-  "version": "7"
-}
